# MAINT: Make the refactor suggested in prepare_index