Split off from #8276 

Extracting the tuple-conversion logic into its own function.

Things I'd like feedback on:

* <s>This might worsen performance slightly for scalar indices, as it now forces the tuple conversion. Is there an easy way to tell if this is significant? Is the clarity vs efficiency tradeoff acceptable here?</s>
* Have I messed up my reference counting?
* Are these functions in the right places?

Update: new approach that:
* Avoids calling `__getitem__` more than once on the provided sequence, if it is converted on a tuple:
  * More performant than `master` for `a[[0, None]]`
* Avoids allocating any new python objects, instead using a tuple-like buffer on the stack