@eric-wieser 2016-11-15
Once this is merged, #4434 will need a rebase