@seberg 2016-12-04
Could be good to see whether it makes a real speed difference to build that tuple (likely not, but then I may actually have timed it at the time and thought it might be nice to have). Then again, I am not sure whether `arr[0]` being even 10-20% slower is a huge deal.