@seberg 2016-12-04
Ref counting looks good on first sight, and yes, definitely all in the right place, the only issue may be to check whether doing a bit weirder code for speed may be worth it.