@eric-wieser 2016-12-04
Ok, all the style things are fixed