@eric-wieser 2016-12-04
> Could be good to see whether it makes a real speed difference to build that tuple (likely not, but then I may actually have timed it at the time and thought it might be nice to have). Then again, I am not sure whether arr[0] being even 10-20% slower is a huge deal.

I think this is basically testable right now as
```python
>>> x = np.arange(1000)
>>> i = 100
>>> %timeit x[i]
The slowest run took 26.21 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 214 ns per loop

>>> %timeit x[(i,)]
The slowest run took 20.53 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 273 ns per loop
```