@seberg 2016-12-04
Hehe, honestly not sure its worth troubling over. If you replace `index_as_tuple = Py_BuildValue("(O)", index);` with lower level calls, you probably get back already a lot of the difference.