@eric-wieser 2016-12-04
`PyTuple_Pack(1, index)` would probably be faster then, and not really any less clear?