@seberg 2016-12-04
Probably, I am not sure, but somewhat thought those pack functions are not the quickest when it comes to micro optimization. Might also be that it does not make a difference at all....