@eric-wieser 2016-12-05
There's a weird uncommented check on [line 190](https://github.com/eric-wieser/numpy/blob/0d5051024bd5436fd21fbbd9112c85244e2f6b7b/numpy/core/src/multiarray/mapping.c#L190) that seems to decide that lists of length MAX_DIM are not promoted to tuples, despite the fact that doing so would be valid up to `2*MAX_DIM`. Is this a backwards compatibility thing? Is this documented somewhere, so that I can add a comment acordingly?