@seberg 2016-12-05
Well the check is only there for the tuple conversion trick, which I am not about to modify, since I think its a crappy hack in any case. Most the time, it is `MAX_DIM` anyway, but since you can have things like `None` which can be there more often, I put `2*MAX_DIM` as basically "for sure enough".