@eric-wieser 2016-12-21
Any more thoughts on this?