@seberg 2017-01-05
Ok, just trying to pass through a few things. Not really, do you know if the slowdown got better now, or do you think we should just not worry about it too much?