@eric-wieser 2017-01-05
I wasn't really able to profile it anyway, being unable to build locally - the closest I could get were:
* `i = 2; %timeit x[i]`
* `ti = (2,); %timeit x[ti]` - profiling of the extra work to iterate the tuple in get_item
* `i = 2; %timeit x[i,]`  - including the worst case tuple constructor cost. A C function call should be less expensive than an opcode though.

These all seemed pretty similar, and the timing overhead seems to make the results pretty hard to compare. So I'd say that we shouldn't worry about it.