@seberg 2017-01-05
Ohh, you have problems building locally? Doesn't the `python runtests.py` and `python runtests.py --ipython` work for you?

OK, just ran it on my computer (python3, on python2 things might be different, because it goes through the C getitem function which directly gets the integer, would have to look up the path myself...).

The new times first:
```python
In [1]: x = np.arange(1000)
In [2]: i = 100
In [3]: %timeit x[i]
10000000 loops, best of 3: 149 ns per loop / 116 ns
In [4]: x = np.arange(1000, dtype=object)
In [5]: %timeit x[i]
10000000 loops, best of 3: 119 ns per loop / 78.2 ns
```

So, hmmmmmm ;). Can't make up my mind, you are right that the difference is lower then doing something like `i+i`.