@eric-wieser 2017-01-05
Wait, why is `dtype=object faster in both cases`?