@seberg 2017-01-05
Simple, it only has to do an incref and not create the scalar python object from whats inside the array.