@seberg 2017-01-11
I suppose this: I think we can get away with it, but would prefer to check whether it is not too ugly if we try to avoid it.