@eric-wieser 2017-04-07
Ok, timing is now much better, and actually an improvement when passing tuples

Before this PR:

```python
In [1]: a = np.array([1, 2, 3, 4])

In [2]: %timeit a[0]
The slowest run took 29.59 times longer than the fastest. This could mean that an intermediate result is being cached.
10000000 loops, best of 3: 158 ns per loop

In [3]: %timeit a[0,]
The slowest run took 28.01 times longer than the fastest. This could mean that an intermediate result is being cached.
10000000 loops, best of 3: 167 ns per loop

```

After the latest commit:

```python
In [1]: a = np.array([1, 2, 3, 4])

In [2]: %timeit a[0]
The slowest run took 32.66 times longer than the fastest. This could mean that an intermediate result is being cached.
10000000 loops, best of 3: 157 ns per loop

In [2]: %timeit a[(0,)]
The slowest run took 26.58 times longer than the fastest. This could mean that an intermediate result is being cached.
10000000 loops, best of 3: 158 ns per loop
```