@seberg 2017-04-08
Probably good, I did not read too carefully though, yet. However, the reference counting seems odd to me, where are the incref's/decref's? As far as I can see, you cannot be sure that anyone else holds a reference to the items you take out of an arbitrary sequence. `GetItem` probably returns a new reference also (which is good here), but there should be a decref when the index parsing is done (in which case also the single item paths probably need increfs).