@eric-wieser 2017-04-08
> However, the reference counting seems odd to me, where are the incref's/decref's

`GET_ITEM` returns a borrow reference, so no reference counting is needed. We just unpack the tuple, and either return a borrowed reference to the tuple itself, or return a borrowed reference to each of its items.

I don't think the reference handling code is any different to what was here originally - in both cases, we borrow everything and increment nothing