@eric-wieser 2017-04-09
Damn, you're very right, and that makes this "manual tuple" a much less reasonable thing to work with