@eric-wieser 2017-04-09
Ok, added reference counting. This is still more performant than master, and definitely has the edge over the same code using normal `PyTuple` cals