@eric-wieser 2017-04-10
@seberg: Let me know if the refcounting now looks good, then I'll squash together the 5 most recent commits