@seberg 2017-07-15
Hehe, right broadcasting of indices.... anyway, yeah, its plenty as is.