@eric-wieser 2017-07-18
Yep, I think squashing via github into one commit is the best plan. The git history is just clutter, but that means if people really care they can check this PR in its unmodified messy state